requirements.txt
fastapi==0.104.1
PyJWT==2.8.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
motor==3.3.2
pymongo==4.6.0
openai-whisper==20231117
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
transformers==4.36.2
google-generativeai==0.3.2
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
email-validator==2.1.0
requests==2.31.0
httpx==0.25.2
# Audio processing essentials
audioread==3.0.1
resampy==0.4.2
numba==0.58.1
# Hugging Face for audio models
datasets==2.14.6
accelerate==0.24.1
# Audio processing alternatives
torch-audiomentations==0.11.0
# Multi-language support
polyglot==16.7.4
pycld2==0.41
langdetect==1.0.9
# Alternative for advanced audio processing
pyannote-audio==3.1.1